Display All Servers From All Manifests in One List
To eliminate the need to switch between manifests to choose a server,
display all the servers from all manifests in one list. When a server is
selected in the list, set the manifest that it came from as the current
manifest. If the selected manifest is the same as the previously
selected one, then no validation is necessary.

Remove the drop-down menu for manifests since it is redundant.

To prevent switching manifests during validation, disable the server
list until validation is complete, or an error occurs.